:hammer: [Fix] user-image 순환참조 수정 #72
