+++ server/src/main/java/com/photoday/photoday/image/service/S3Service.java
@@ -1,0 +1,36 @@
+package com.photoday.photoday.image.service;
+
+import com.amazonaws.services.s3.AmazonS3Client;
+import com.amazonaws.services.s3.model.ObjectMetadata;
+import com.amazonaws.services.s3.model.PutObjectRequest;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.IOException;
+
+@Service
+@RequiredArgsConstructor
+public class S3Service {
+    private final AmazonS3Client amazonS3Client;
+    @Value("${cloud.aws.s3.bucket}")
+    private String s3Bucket;
+
+    public String saveImage(MultipartFile multipartFile) throws IOException {
+        String originalFilename = multipartFile.getOriginalFilename();
+        long size = multipartFile.getSize();
+
+        ObjectMetadata objectMetadata = new ObjectMetadata();
+        objectMetadata.setContentType(multipartFile.getContentType());
+        objectMetadata.setContentLength(size);
+
+        //TODO 메타데이터 저장, pre signed url 고려
+
+        amazonS3Client.putObject(
+                new PutObjectRequest(s3Bucket, originalFilename, multipartFile.getInputStream(), objectMetadata)
+        );
+
+        return amazonS3Client.getUrl(s3Bucket, originalFilename).toString();
+    }
+}
